mesh data classes written with generic storage class. changing some import libraries. Starting to refactor the box example to use new mesh classes as a test for this implementation

+++ DX12Demo/IndexBufferExample.cpp
@@ -55,7 +55,7 @@ namespace BoulderLeaf::Graphics::DX12
 		);
 
 		mBufferView.BufferLocation = mVertexBufferGPU->GetGPUVirtualAddress();
-		mBufferView.SizeInBytes = vbByteSize;
+		mBufferView.SizeInBytes = static_cast<UINT>(vbByteSize);
 		mBufferView.StrideInBytes = sizeof(Vertex);
 
 		mBufferViews[1] = { mBufferView };
